Feat/#36

+++ Mindspace_back/src/comment/comment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { PagingParams } from '../global/common/type';
@@ -10,6 +10,10 @@ import { CommentResponseDto } from './dto/comment-response.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { BoardService } from '../board/board.service';
 import { CustomCommentRepository } from './repository/comment.repository';
+import { CommentPermissionDeniedException } from './exception/CommentPermissionDeniedException';
+import { CommentNotFoundException } from './exception/CommentNotFoundException';
+import { UserNotFoundException } from '../user/exception/UserNotFoundException';
+import { BoardNotFoundException } from '../board/exception/BoardNotFoundException';
 
 @Injectable()
 export class CommentService {
@@ -22,6 +26,7 @@ export class CommentService {
     private readonly boardService: BoardService,
   ) {}
 
+  /** 댓글 생성 */
   async createComment(
     boardId: number,
     userId: string,
@@ -31,13 +36,13 @@ export class CommentService {
     const user = await this.userService.findUserById(convertedUserId);
 
     if (!user) {
-      throw new NotFoundException('사용자를 찾을 수 없습니다.');
+      throw new UserNotFoundException();
     }
 
     const board = await this.boardService.findBoardById(boardId);
 
     if (!board) {
-      throw new NotFoundException('게시글을 찾을 수 없습니다.');
+      throw new BoardNotFoundException();
     }
 
     const comment = this.commentMapper.DtoToEntity(
@@ -48,7 +53,26 @@ export class CommentService {
     );
     return await this.commentRepository.save(comment);
   }
-  async getCommentsByBoardId(boardId: number, pagingParams: PagingParams) {
+
+  /** 댓글 목록 조회 */
+  async getCommentsByBoardId(
+    boardId: number,
+    userId: string,
+    pagingParams: PagingParams,
+  ) {
+    const convertedUserId = Number(userId);
+    const user = await this.userService.findUserById(convertedUserId);
+
+    if (!user) {
+      throw new UserNotFoundException();
+    }
+
+    const board = await this.boardService.findBoardById(boardId);
+
+    if (!board) {
+      throw new BoardNotFoundException();
+    }
+
     const comments = await this.customCommentRepository.paginate(
       boardId,
       pagingParams,
@@ -56,7 +80,7 @@ export class CommentService {
 
     const transformedComments: CommentResponseDto[] = comments.data.map(
       (comment) => {
-        return CommentMapper.commentToResponseDto(comment);
+        return CommentMapper.commentToResponseDto(comment, userId);
       },
     );
 
@@ -66,27 +90,58 @@ export class CommentService {
     };
   }
 
+  /** 댓글 수정 */
   async updateComment(
     comment_id: number,
+    userId: string,
     updateCommentDto: UpdateCommentDto,
   ): Promise<UpdateCommentDto> {
+    const convertedUserId = Number(userId);
+    const user = await this.userService.findUserById(convertedUserId);
+
+    if (!user) {
+      throw new UserNotFoundException();
+    }
+
     const comment = await this.commentRepository.findOne({
       where: { id: comment_id },
+      relations: ['user'],
     });
+
     if (!comment) {
-      throw new Error('댓글을 찾을 수 없습니다.');
+      throw new CommentNotFoundException();
     }
+
+    if (comment.user.id.toString() !== userId) {
+      throw new CommentPermissionDeniedException();
+    }
+
     comment.content = updateCommentDto.content;
     return await this.commentRepository.save(comment);
   }
 
-  async deleteComment(comment_id: number): Promise<void> {
+  /** 댓글 삭제 */
+  async deleteComment(comment_id: number, userId: string): Promise<void> {
+    const convertedUserId = Number(userId);
+    const user = await this.userService.findUserById(convertedUserId);
+
+    if (!user) {
+      throw new UserNotFoundException();
+    }
+
     const comment = await this.commentRepository.findOne({
       where: { id: comment_id },
+      relations: ['user'],
     });
+
     if (!comment) {
-      throw new Error('댓글을 찾을 수 없습니다.');
+      throw new CommentNotFoundException();
     }
+
+    if (comment.user.id.toString() !== userId) {
+      throw new CommentPermissionDeniedException();
+    }
+
     await this.commentRepository.remove(comment);
   }
 }
